fix(libminiooni): follow conventions of other packages
All packages use `foo_test.go` and `foo_integration_test.go`, so
do the same also for the libminiooni package.

See https://github.com/ooni/probe-engine/issues/1005
