+++ libminiooni/libminiooni_integration_test.go
@@ -1,0 +1,13 @@
+// +build integration
+
+package libminiooni_test
+
+import (
+	"testing"
+
+	"github.com/ooni/probe-engine/libminiooni"
+)
+
+func TestIntegrationSimple(t *testing.T) {
+	libminiooni.MainWithConfiguration("example", libminiooni.Options{})
+}
